refactor: isolate workshop revision handoff
