+++ ui/src/app/bootstrap.ts
@@ -32,7 +32,6 @@ import type {
   ApplicationContext,
   ApplicationNavigationPreferences,
   ApplicationNavigationPreferencesSnapshot,
-  ApplicationSkillWorkshopRevisionHandoff,
   ApplicationTheme,
   ApplicationThemeServerSelection,
 } from "./context.ts";
@@ -51,6 +50,7 @@ import {
   saveSettings,
   type UiSettings,
 } from "./settings.ts";
+import { createSkillWorkshopRevisionHandoff } from "./skill-workshop-revision-handoff.ts";
 import { createStartupLifecycle, type StartupStep } from "./startup-lifecycle.ts";
 import { resolveApplicationStartupSettings } from "./startup-settings.ts";
 import { startThemeTransition } from "./theme-transition.ts";
@@ -200,28 +200,6 @@ function createApplicationNavigationPreferences(
     subscribe(listener) {
       listeners.add(listener);
       return () => listeners.delete(listener);
-    },
-  };
-}
-
-export function createSkillWorkshopRevisionHandoff(): ApplicationSkillWorkshopRevisionHandoff {
-  let pending: Parameters<ApplicationSkillWorkshopRevisionHandoff["prepare"]>[0] | null = null;
-  return {
-    prepare: (handoff) => {
-      pending = handoff;
-    },
-    consume: (sessionKey, owner) => {
-      if (!pending || pending.sessionKey !== sessionKey || pending.owner !== owner) {
-        return null;
-      }
-      const handoff = pending;
-      pending = null;
-      return handoff;
-    },
-    clear: (handoff) => {
-      if (!handoff || pending === handoff) {
-        pending = null;
-      }
     },
   };
 }
